requirements-dev.txt
robotframework >= 3.0
robotstatuschecker == 1.5.1
robotbackgroundlogger >= 1.2
